% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsbart_regression.R
\name{gsbart_regression}
\alias{gsbart_regression}
\title{General SoftBart Regression with group handling}
\usage{
gsbart_regression(
  X,
  Y,
  Z,
  Znew = Z,
  Xnew = X,
  num_tree = 50,
  k = 2,
  alpha_val = 1,
  alpha_comp = 1,
  alpha_scale = NULL,
  alpha_comp_scale = NULL,
  num_burn = 5000,
  num_thin = 10,
  num_save = 500,
  hypers = NULL,
  opts = NULL,
  group = NULL,
  include_intercept = FALSE
)
}
\arguments{
\item{X}{A matrix of linear covariate included in the mean response.}

\item{Y}{A vector of outcome data.}

\item{Z}{A matrix of predictor variables to be included in the exposure-response function.}

\item{Znew}{Optional matrix of new predictor values.}

\item{Xnew}{Optional matrix of new covariate values.}

\item{num_tree}{The number of trees in the ensemble to use.}

\item{k}{Determines the standard deviation of the leaf node parameters, which is given by \code{3 / k / sqrt(num_tree)}.}

\item{alpha_val}{Positive constant controlling the sparsity level among groups.}

\item{alpha_comp}{Positive constant controlling the sparsity level across group components.}

\item{alpha_scale}{Scale of the prior for \code{alpha}.}

\item{alpha_comp_scale}{Scale of the prior for \code{alpha_comp}.}

\item{num_burn}{Number of warmup iterations for the chain.}

\item{num_thin}{Thinning interval for the chain.}

\item{num_save}{The number of samples to collect; in total, \code{num_burn + num_save * num_thin} iterations are run.}

\item{hypers}{A list of hyperparameters constructed from the \code{Hypers()} function (\code{num_tree}, \code{k}, \code{alpha_val}, \code{alpha_comp}, \code{alpha_scale} and \code{alpha_comp_scale} are overridden by this function).}

\item{opts}{A list of options for running the chain constructed from the \code{Opts()} function (\code{num_burn}, \code{num_thin}, and \code{num_save} are overridden by this function).}

\item{group}{Optional vector of group indicators used for hierarchical variable selection. If not specified, variable selection is performed at the component level.}

\item{include_intercept}{If \code{FALSE} then any intercept term in the linear formula will be removed, with the overall location of the outcome captured by the nonparametric function.}
}
\value{
Returns a list with the following components
\itemize{
  \item \code{h_train}: samples of the nonparametric function evaluated on the training set.
  \item \code{h_test}: samples of the nonparametric function evaluated on the test set.
  \item \code{eta_train}: samples of the linear predictor on the training set.
  \item \code{eta_test}: samples of the linear predictor on the test set.
  \item \code{beta_hat}: samples of the regression coefficients.
  \item \code{sigma}: samples of the error standard deviation.
  \item \code{sigma_mu}: samples of the standard deviation of the leaf node parameters.
  \item \code{var_counts}: a matrix with a column for each nonparametric predictor containing the number of times that predictor is used in the ensemble at each iteration.
  \item \code{tree_counts}: a list of counts at tree level. If \code{group} is \code{NULL}, each element is a matrix with one column per tree and one row per nonparametric predictor, giving the number of times each predictor is used in each tree. If \code{group} is specified, each element contains two matrices: one with group counts by tree and one with predictor counts by tree.
  \item \code{opts}: the options used when running the chain.
  \item \code{hypers}: the hypers used when running the chain.
}
}
\description{
Fits a semiparametric regression model with grouped predictors and linear covariates using a SoftBart model.
}
